Feature LWDEV-6465 add timestamp check

Approved-by: Oleg Zhelenkov <oleg.zhelenkov@lykkex.com>

+++ src/Lykke.Job.ForwardWithdrawalResolver/Sagas/CommandsHandler.cs
@@ -71,39 +71,66 @@ namespace Lykke.Job.ForwardWithdrawalResolver.Sagas
         [UsedImplicitly]
         public async Task<CommandHandlingResult> Handle(RemoveEntryFromHistoryServiceCommand command, IEventPublisher eventPublisher)
         {
-            _log.WriteInfo(nameof(RemoveEntryFromHistoryServiceCommand), command.ClientId, $"Beginning to process: {command.ToJson()}");
-
-            await _operationsHistoryClient.DeleteByClientIdOperationId(command.ClientId, command.CashInId);
-            
-            eventPublisher.PublishEvent(new CashInRemovedFromHistoryServiceEvent
+            try
             {
-                Id = command.Id,
-                ClientId = command.ClientId,
-                AssetId = command.AssetId,
-                Amount = command.Amount,
-                CashInId = command.CashInId
-            });
-            
-            return CommandHandlingResult.Ok();
+                _log.WriteInfo(nameof(RemoveEntryFromHistoryServiceCommand), command.ClientId,
+                    $"Beginning to process: {command.ToJson()}");
+                
+                if(command.CashInId == null)
+                    throw new InvalidOperationException($"CashInId of {command.Id} FW should not be null");
+
+                await _operationsHistoryClient.DeleteByClientIdOperationId(command.ClientId, command.CashInId);
+
+                eventPublisher.PublishEvent(new CashInRemovedFromHistoryServiceEvent
+                {
+                    Id = command.Id,
+                    ClientId = command.ClientId,
+                    AssetId = command.AssetId,
+                    Amount = command.Amount,
+                    CashInId = command.CashInId
+                });
+
+                return CommandHandlingResult.Ok();
+            }
+            catch (Exception e)
+            {
+                _log.WriteError(nameof(RemoveEntryFromHistoryServiceCommand), command.ClientId, e);
+                
+                return CommandHandlingResult.Fail(TimeSpan.FromSeconds(30));
+            }
         }
-        
-        [UsedImplicitly]
-        public async Task<CommandHandlingResult> Handle(RemoveEntryFromHistoryJobCommand command, IEventPublisher eventPublisher)
-        {
-            _log.WriteInfo(nameof(RemoveEntryFromHistoryJobCommand), command.ClientId, $"Beginning to process: {command.ToJson()}");
 
-            await _operationsCacheClient.RemoveCashInIfExists(command.ClientId, command.CashInId);
-            
-            eventPublisher.PublishEvent(new CashInRemovedFromHistoryJobEvent
+        [UsedImplicitly]
+        public async Task<CommandHandlingResult> Handle(RemoveEntryFromHistoryJobCommand command,
+            IEventPublisher eventPublisher)
+        {
+            try
             {
-                Id = command.Id,
-                ClientId = command.ClientId,
-                AssetId = command.AssetId,
-                Amount = command.Amount,
-                CashInId = command.CashInId
-            });
-            
-            return CommandHandlingResult.Ok();
+                _log.WriteInfo(nameof(RemoveEntryFromHistoryJobCommand), command.ClientId,
+                    $"Beginning to process: {command.ToJson()}");
+
+                if (command.CashInId == null)
+                    throw new InvalidOperationException($"CashInId of {command.Id} FW should not be null");
+
+                await _operationsCacheClient.RemoveCashInIfExists(command.ClientId, command.CashInId);
+
+                eventPublisher.PublishEvent(new CashInRemovedFromHistoryJobEvent
+                {
+                    Id = command.Id,
+                    ClientId = command.ClientId,
+                    AssetId = command.AssetId,
+                    Amount = command.Amount,
+                    CashInId = command.CashInId
+                });
+
+                return CommandHandlingResult.Ok();
+            }
+            catch (Exception e)
+            {
+                _log.WriteError(nameof(RemoveEntryFromHistoryJobCommand), command.ClientId, e);
+
+                return CommandHandlingResult.Fail(TimeSpan.FromSeconds(30));
+            }
         }
 
         [UsedImplicitly]
